+++ sw/airborne/modules/computer_vision/farneback/farneback_navigator.h
