+++ sw/airborne/modules/computer_vision/farneback/finite_state_machine.c
